+++ src/assets/style.scss
@@ -11,28 +11,33 @@
 
 /******* Section *******/
 .section__container {
-  background-color: #F4F5F7;
+  background-color: #E6E8ED;
   display: flex;
   flex-flow: row wrap;
   width: 100%;
   height: 100%;
   min-height: 100vh;
-  padding: 80px 0 50px 0;
+  padding: 54px 0 50px 0;
   box-sizing: border-box;
 }
 
 .home__center {
   width: calc((100% - 80px) * 0.6);
+  padding-top: 30px;
   margin-left: 20px;
 }
 
 .home__left {
   width: calc((100% - 80px) * 0.2);
+  padding-top: 30px;
+
   margin-left: 20px;
 }
 
 .home__right {
   width: calc((100% - 80px) * 0.2);
+  padding-top: 30px;
+
   margin: 0 20px;
 }
