+++ src/assets/todo/switch-today-or-monthly-button.scss
@@ -1,7 +1,7 @@
 @import "../variable";
 @import "../mixin/main-btn";
 
-.switch-date-button {
+.switch-today-or-monthly-button {
   display: flex;
   height: 100%;
