+++ src/assets/todo/todo.scss
@@ -3,58 +3,46 @@
 
 .todo {
   width: 100%;
-  display: flex;
 
-  &__today-list {
-    width: calc((100% - 60px) * 0.7);
-    margin-left: 20px;
-  }
-
-  &__today-list-content {
-    background-color: white;
-    border-radius: 0.6rem;
-    padding: 20px;
-  }
-
-  &__expired-list {
-    width: calc((100% - 60px) * 0.3);
-    margin: 0 20px;
-  }
-
-  &__expired-list-content {
-    background-color: white;
-    border-radius: 0.6rem;
-    padding: 15px;
-  }
-
-  &__menu {
+  &__switch-page {
     display: flex;
-    margin-bottom: 30px;
+    width: 100%;
+    background-color: white;
+    text-align: center;
+    margin-bottom: 40px;
   }
 
-  &__search {
-    @extend .task--btn;
-    height: 40px;
-    margin-left: 20px;
+  &__switch-page-button {
+    width: 50%;
+    background-color: white;
+    border: none;
+    color: #5D6B89;
+    font-size: 0.9rem;
+    padding: 10px 18px 5px 18px;
+    cursor: pointer;
+
+    &:focus {
+      outline: none;
+    }
   }
 
-  &__today-date {
-    display: inline-block;
-    color: $task-main_color;
-    font-size: 1.2rem;
+  &__current-page-button {
+    width: 50%;
+    border: none;
+    background-color: #E6E8ED;
+    color: #515C77;
+    font-size: 0.9rem;
     font-weight: bold;
-    margin: 0 0 10px 10px;
-  }
+    padding: 10px 18px 5px 18px;
+    cursor: pointer;
 
-  &__today-switch-todo-list {
-    background-color: #F4F5F7;
-    border-radius: 0.6rem;
-    padding: 25px 0;
-    height: auto;
+    &:focus {
+      outline: none;
+    }
 
-    &--width {
-      width: 60%;
-      margin: auto;
+    > span {
+       border-bottom: #F56E6B 3px solid;
+       padding-bottom: 6px;
     }
   }
 }
