+++ src/components/todo/index.tsx
@@ -5,7 +5,7 @@ export { default as SearchField } from './SearchField';
 export { default as SearchTodoList } from './SearchTodoList';
 export { default as SearchResultTodoList } from './SearchResultTodoList';
 export { default as SelectDateType } from './SelectDateType';
-export { default as SwitchDateButton } from './SwitchDateButton';
+export { default as SwitchDateButton } from './SwitchTodayOrMonthlyButton';
 export { default as TextInput } from './TextInput';
 export { default as TodoButton } from './TodoButton';
 export { default as TodoListItemComponent } from './TodoListItemComponent';
