+++ src/components/todo/todo-page/todo-page.scss
@@ -1,0 +1,60 @@
+@import "../../../assets/variable";
+@import "../../../assets/modules/task-btn";
+
+.todo-page {
+  width: 100%;
+  display: flex;
+
+  &__today-list {
+    width: calc((100% - 60px) * 0.7);
+    margin-left: 20px;
+  }
+
+  &__today-list-content {
+    background-color: white;
+    border-radius: 0.6rem;
+    padding: 20px;
+  }
+
+  &__expired-list {
+    width: calc((100% - 60px) * 0.3);
+    margin: 0 20px;
+  }
+
+  &__expired-list-content {
+    background-color: white;
+    border-radius: 0.6rem;
+    padding: 15px;
+  }
+
+  &__menu {
+    display: flex;
+    margin-bottom: 30px;
+  }
+
+  &__search {
+    @extend .task--btn;
+    height: 40px;
+    margin-left: 20px;
+  }
+
+  &__today-date {
+    display: inline-block;
+    color: $task-main_color;
+    font-size: 1.2rem;
+    font-weight: bold;
+    margin: 0 0 10px 10px;
+  }
+
+  &__today-switch-todo-list {
+    background-color: #F4F5F7;
+    border-radius: 0.6rem;
+    padding: 25px 0;
+    height: auto;
+
+    &--width {
+      width: 60%;
+      margin: auto;
+    }
+  }
+}
